+++ R/add_boundary_penalties.R
@@ -129,8 +129,10 @@ NULL
 #'
 #' @seealso
 #' See [penalties] for an overview of all functions for adding penalties.
-#' Additionally, see [add_neighbor_penalties()] for a penalty that
+#' Also see [add_neighbor_penalties()] for a penalty that
 #' can reduce spatial fragmentation and has faster solver run times.
+#' Additionally, see [calibrate_cohon_method()] for assistance with selecting
+#' an appropriate `penalty` value.
 #'
 #' @family penalties
 #'