+++ R/add_connectivity_penalties.R
@@ -146,8 +146,10 @@ NULL
 #'
 #' @seealso
 #' See [penalties] for an overview of all functions for adding penalties.
-#' Additionally, see [add_asym_connectivity_penalties()] to account for
+#' Also see [add_asym_connectivity_penalties()] to account for
 #' asymmetric connectivity between planning units.
+#' Additionally, see [calibrate_cohon_method()] for assistance with selecting
+#' an appropriate `penalty` value.
 #'
 #' @family penalties
 #'