+++ R/add_linear_penalties.R
@@ -107,6 +107,8 @@ NULL
 #'
 #' @seealso
 #' See [penalties] for an overview of all functions for adding penalties.
+#' Also, see [calibrate_cohon_method()] for assistance with selecting
+#' an appropriate `penalty` value.
 #'
 #' @family penalties
 #'